Create a map based on GaoDe API and Parcel
To record the house information and display it on the map,
use GaoDe API to show the map around Dalian Library.

This commit adds the following functionality:
- Create a map based on GaoDe API
- Add a marker to the map
- Add a popup to the marker

The bad news is that for now the map marker stores the information
in an array, which will lose the information when the page is refreshed.

The next step is to store the information in the local storage.

Signed-off-by: Daliuding <liujing522@163.com>